requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
fastapi==0.95.0
h11==0.14.0
idna==3.4
motor==3.1.2
oauthlib==3.2.2
pydantic==1.10.7
pymongo==4.3.3
python-decouple==3.8
pyzaim==1.0.11
requests==2.28.2
requests-oauthlib==1.3.1
selenium==3.141.0
sniffio==1.3.0
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
